ecommerce_backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
django-filter==25.2
psycopg2-binary==2.9.11
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
python-dotenv==1.0.1
